+++ app/components/WebsiteCard.vue
@@ -1,14 +1,22 @@
 <script>
+import { recordClick } from '~/composables/useClickTracker'
+
 export default {
   props: {
     websiteItem: { type: Object, required: true }
+  },
+  methods: {
+    trackClick() {
+      recordClick(this.websiteItem.href)
+      this.$emit('tracked')
+    }
   }
 }
 </script>
 
 <template>
   <li>
-    <NuxtLink :to="websiteItem.href" :title="websiteItem.desc" target="_blank" rel="noopener noreferrer" class="card-link">
+    <NuxtLink :to="websiteItem.href" :title="websiteItem.desc" target="_blank" rel="noopener noreferrer" class="card-link" @click="trackClick">
       <div class="card-icon">
         <span v-if="websiteItem.logo === ''" class="card-avatar">{{ websiteItem.name.charAt(0) }}</span>
         <img v-else :src="websiteItem.logo" :alt="websiteItem.name" loading="lazy" class="card-logo">